Altered make_move to allow string params in SAN.
The original implementation of make_move required that the moves be
passed in move object format.  While this is effecient, it's not very
programmer friendly.  make_move now supports as a paramters a move
object or a string in standard algebraic notation (e.g.
chess.make_move('Qh5+')).  A note, algebraic notation is case sensitive
and must contain the appropriate decorators (e.g. + for check, # for
mate, etc.).