Optimisations for history, pgn and getComments
Improved the efficiency of several methods by including the pretty move in _history. Therefore, the entire game does not need to be replayed whenever the history needs to be iterated just to get the fen/san moves in each position.

Added an optional { pushPretty: false } flag to _makeMove to adjust its behaviour. When provided, the pretty move will not be pushed to the history. This is so that when _makeMove is called only to be immediately followed by an _undoMove (which is a common design pattern throughout the library), the pretty move is not generated unnecessarily (and indeed potentially causing infinite recursion). However, this flag should only be used when the move is not intended to be left on the history.